net/ixgbe: fix queue received bytes CRC adjustment
For 82599, QBRC is a 36-bit clear-on-read counter, while
QPRC is a 32-bit clear-on-read counter. ixgbe_read_stats_registers()
accumulates QBRC in a 64-bit software counter and, when CRC stripping
is disabled, subtracts the CRC bytes accounted for each received packet.

The CRC adjustment is computed as:

	delta_qprc * RTE_ETHER_CRC_LEN

Since delta_qprc is 32-bit, the multiplication is performed in 32 bits
and may wrap before the result is subtracted from the 64-bit QBRC
accumulator. A full 32-bit packet delta needs more than 32 bits to
represent the CRC-byte adjustment.

Cast delta_qprc to uint64_t before the multiplication so the adjustment
is computed with the same effective width as the byte counter.

Found by Linux Verification Center (linuxtesting.org) with SVACE.

Fixes: c03fcee9abbd ("ixgbe: remove CRC size from byte counters")
Cc: stable@dpdk.org

Acked-by: Bruce Richardson <bruce.richardson@intel.com>

+++ drivers/net/intel/ixgbe/ixgbe_ethdev.c
@@ -3246,7 +3246,7 @@ ixgbe_read_stats_registers(struct ixgbe_hw *hw,
 		hw_stats->qbrc[i] +=
 		    ((uint64_t)IXGBE_READ_REG(hw, IXGBE_QBRC_H(i)) << 32);
 		if (crc_strip == 0)
-			hw_stats->qbrc[i] -= delta_qprc * RTE_ETHER_CRC_LEN;
+			hw_stats->qbrc[i] -= (uint64_t)delta_qprc * RTE_ETHER_CRC_LEN;
 
 		hw_stats->qbtc[i] += IXGBE_READ_REG(hw, IXGBE_QBTC_L(i));
 		hw_stats->qbtc[i] +=
